feat: Add GameEnd and NewGame components
This commit adds two new components, GameEnd and NewGame, to the game module. The GameEnd component displays a message when the game ends, indicating whether it is a checkmate or a stalemate. The NewGame component allows the user to start a new game. These components are used in the GamePage component to provide a better user experience.